requirements.txt
asgiref
crispy-bootstrap5
distlib
dj-database-url
Django
django-cors-headers
django-crispy-for
django-email
django-environ
django-filte
django-phone-field
django-phonenumber-field
djangorestframework
env
envir
filelock
Library
lxml
Markdo
MouseInfo
numpy
opencv-contrib-python
pillow
platformdirs
psycopg2
psycopg2-binary
PyAutoGUI
PyGetWindow
PyMsgBox
pyperclip
PyRect
PyScreeze
python-docx
pytweening
sqlparse
typing_extensions
tzdata
virtualenv
